requirements.txt
streamlit>=1.28.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
numpy>=1.24.0
google-generativeai>=0.3.0
pdf2image>=1.16.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
subprocess32; python_version<"3.2"
